+++ CHANGELOG.md
@@ -18,6 +18,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Dropped support for legacy index files written by jj < 0.33. New index files
   will be created as needed.
 
+* The following deprecated config options have been removed:
+  - `core.fsmonitor`
+  - `core.watchman.register-snapshot-trigger`
+
 ### Deprecations
 
 * `jj debug snapshot` is deprecated in favor of `jj util snapshot`. Although